refactor(typography): rename .scss files to .module.scss
