+++ packages/typography/src/Heading/Heading.tsx
@@ -1,7 +1,7 @@
 import { createElement, HTMLAttributes } from "react"
 import classnames from "classnames"
 import { OverrideClassName } from "@kaizen/component-base"
-import styles from "./Heading.scss"
+import styles from "./Heading.module.scss"
 
 const VARIANTS_24PX_OR_GREATER = ["display-0", "heading-1", "heading-2"]
