+++ packages/typography/src/Paragraph/Paragraph.tsx
@@ -1,7 +1,7 @@
 import { createElement, HTMLAttributes } from "react"
 import classnames from "classnames"
 import { OverrideClassName } from "@kaizen/component-base"
-import styles from "./Paragraph.scss"
+import styles from "./Paragraph.module.scss"
 
 export type ParagraphVariants = "intro-lede" | "body" | "small" | "extra-small"
